semi complete terminal integration
have mostly finished terminal integrations i just have a few more things to implement properly before its fully finished and i am also going to go through and make some updates to some of the text in the game itself to make sure that the player knows whats going on at all times. i also want to be able to start working on a statistics implementation within the terminal script so that when in combat or when the player opens their menu a little side pannel will pop up showing statistics like health, level, combat EXP, healing items left and maybe even how much exp till the next level kind of things. tho this is still in planning and might not be implemented untill later on.